requirements.txt
Requests==2.31.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.12
